# src/batched/dense: Add Gemm_DblBuf LayoutLeft operator