This PR is based off of the source branch in #1256. #1256 should be merged before reviewing these changes.

- New operator indexes global and scratch memory such that accesses across vlanes are contiguous.

## Performance improvement (double scalar type, batch size of 8192)
square matrix dims|Before (Gflop/s)|After (Gflop/s)
| ----------- | ----------- | ----------- |
2|16.8544|17.6915
4|135.402|141.826
6|170.94|169.095
8|265.219|266.223
10|283.863|272.781
12|398.604|400.089
14|327.058|331.717
16|391.265|499.418
18|430.532|562.897
20|543.968|682.534
22|649.959|780.68
24|798.063|986.864
26|780.774|947.36
28|928.002|1139.47
30|1031.92|1227.36
32|1446.63|1538.47
34|605.81|739.448
36|679.656|838.312
38|758.232|936.194
40|839.289|1048.91
42|887.601|1098.1
44|974.299|1211.74
46|1053.33|1321.66
48|1162.6|1478.52
50|1202.08|1489.85
52|1312.51|1617.37
54|1437.34|1731.18
56|1546.69|1908.75
58|1556.72|1887.8
60|1690.83|2058.28
62|1788.88|2150.14
64|2394.12|2862.58

Fixes #1274.